Merging LL updates into rk-lindsey

+++ src/chimes_run_md.py
@@ -42,8 +42,8 @@ def post_proc(my_ALC, my_case, my_indep, *argv, **kwargs):
     
     ### ...kwargs
     
-    default_keys   = [""]*11
-    default_values = [""]*11
+    default_keys   = [""]*12
+    default_values = [""]*12
 
 
     # MD specific controls
@@ -62,6 +62,7 @@ def post_proc(my_ALC, my_case, my_indep, *argv, **kwargs):
     default_keys[8 ] = "loose_crit"  ; default_values[8 ] = "../../../../loose_bond_crit.dat"                            # File with loose bonding criteria for clustering
     default_keys[9 ] = "clu_code"     ; default_values[9 ] = "/p/lscratchrza/rlindsey/RC4B_RAG/11-12-18/new_ts_clu.cpp"  # Clustering code
     default_keys[10] = "compilation" ; default_values[10] = "g++ -std=c++11 -O3"                                         # Command to compile clustering code
+    default_keys[11] = "run_molanal" ; default_values[11] = False              
     
     args = dict(list(zip(default_keys, default_values)))
     args.update(kwargs)
@@ -77,17 +78,19 @@ def post_proc(my_ALC, my_case, my_indep, *argv, **kwargs):
     ################################
     # 1. Run molanal
     ################################
-    
-    if os.path.isfile(args["basefile_dir"] + "case-" + str(my_case) + ".skip.dat"):
-    
-        helpers.run_bash_cmnd("cp " + args["basefile_dir"] + "case-" + str(my_case) + ".skip.dat skip.dat")
-    
-    
-    helpers.run_bash_cmnd_to_file("traj.gen-molanal.out",args["molanal_dir"] + "/molanal.new traj.gen")
-    helpers.run_bash_cmnd_to_file("traj.gen-find_molecs.out", args["molanal_dir"] + "/findmolecules.pl traj.gen-molanal.out")
-    helpers.run_bash_cmnd("rm -rf molecules " + ' '.join(glob.glob("molanal*")))
-    
-    print(helpers.run_bash_cmnd_presplit([args["local_python"], args["driver_dir"] + "/src/post_process_molanal.py"] + args_species))
+    if args["run_molanal"]:
+        if os.path.isfile(args["basefile_dir"] + "case-" + str(my_case) + ".skip.dat"):
+
+            helpers.run_bash_cmnd("cp " + args["basefile_dir"] + "case-" + str(my_case) + ".skip.dat skip.dat")
+
+
+        helpers.run_bash_cmnd_to_file("traj.gen-molanal.out",args["molanal_dir"] + "/molanal.new traj.gen")
+        helpers.run_bash_cmnd_to_file("traj.gen-find_molecs.out", args["molanal_dir"] + "/findmolecules.pl traj.gen-molanal.out")
+        helpers.run_bash_cmnd("rm -rf molecules " + ' '.join(glob.glob("molanal*")))
+
+        print(helpers.run_bash_cmnd_presplit([args["local_python"], args["driver_dir"] + "/src/post_process_molanal.py"] + args_species))
+    else:
+        print("Skipping MOLANAL")
     
     ################################
     # 2. Cluster
@@ -141,20 +144,20 @@ def run_md(my_ALC, my_case, my_indep, *argv, **kwargs):
     
     ### ...kwargs
     
-    default_keys   = [""]*15
-    default_values = [""]*15
+    default_keys   = [""]*16
+    default_values = [""]*16
 
     # MD specific controls
 
     default_keys[0 ] = "basefile_dir"  ; default_values[0 ] = "../CHIMES-MD_BASEFILES/"    # Directory containing run_md.base, etc.
     default_keys[1 ] = "driver_dir"    ; default_values[1 ] = ""                           # Post_proc_lsq*py file... should also include the python command
-    default_keys[2 ] = "penalty_pref"  ; default_values[2 ] = 1.0E6                # Penalty function pre-factor
-    default_keys[3 ] = "penalty_dist"  ; default_values[3 ] = 0.02                 # Pentaly function kick-in distance
-    default_keys[4 ] = "chimes_exe  "  ; default_values[4 ] = None                 # Unused by this function
+    default_keys[2 ] = "penalty_pref"  ; default_values[2 ] = 1.0E6                        # Penalty function pre-factor
+    default_keys[3 ] = "penalty_dist"  ; default_values[3 ] = 0.02                         # Pentaly function kick-in distance
+    default_keys[4 ] = "chimes_exe  "  ; default_values[4 ] = None                         # Unused by this function
     
     # Overall job controls
 
-    default_keys[5 ] = "job_name"       ; default_values[5 ] = "ALC-"+ repr(my_ALC)+"-md"    # Name for ChIMES md job
+    default_keys[5 ] = "job_name"      ; default_values[5 ] = "ALC-"+ repr(my_ALC)+"-md"     # Name for ChIMES md job
     default_keys[6 ] = "job_nodes"     ; default_values[6 ] = "2"                            # Number of nodes for ChIMES md job
     default_keys[7 ] = "job_ppn"       ; default_values[7 ] = "36"                           # Number of processors per node for ChIMES md job
     default_keys[8 ] = "job_walltime"  ; default_values[8 ] = "1"                            # Walltime in hours for ChIMES md job
@@ -162,8 +165,10 @@ def run_md(my_ALC, my_case, my_indep, *argv, **kwargs):
     default_keys[10] = "job_account"   ; default_values[10] = "pbronze"                      # Account for ChIMES md job
     default_keys[11] = "job_executable"; default_values[11] = ""                             # Full path to executable for ChIMES md job
     default_keys[12] = "job_system"    ; default_values[12] = "slurm"                        # slurm or torque    
-    default_keys[13] = "job_file"       ; default_values[13] = "run.cmd"                     # Name of the resulting submit script
+    default_keys[13] = "job_file"      ; default_values[13] = "run.cmd"                      # Name of the resulting submit script
     default_keys[14] = "job_email"     ; default_values[14] = True                           # Send slurm emails?
+    default_keys[15] = "job_modules"   ; default_values[15] = ""                             # Send slurm emails?
+
 
     args = dict(list(zip(default_keys, default_values)))
     args.update(kwargs)    
@@ -272,8 +277,10 @@ def run_md(my_ALC, my_case, my_indep, *argv, **kwargs):
     
     job_task  = "-n " + repr(int(args["job_nodes"])*int(args["job_ppn"])) + " " +  args["job_executable"] + " " + md_infile + " > run_md.out"    
     
-    if args["job_system"] == "slurm":
+    if (args["job_system"] == "slurm" or args["job_system"] == "UM-ARC"):
         job_task = "srun "   + job_task
+    elif args["job_system"] == "TACC":
+        job_task = "ibrun "  + job_task
     else:
         job_task = "mpirun " + job_task    
     
@@ -283,10 +290,11 @@ def run_md(my_ALC, my_case, my_indep, *argv, **kwargs):
         job_nodes      = str(args["job_nodes"   ]),
         job_ppn        = str(args["job_ppn"     ]),
         job_walltime   = str(args["job_walltime"]),
-        job_queue      =     args["job_queue"    ] ,
+        job_queue      =     args["job_queue"   ] ,
         job_account    =     args["job_account" ] ,
         job_executable =     job_task,
         job_system     =     args["job_system"  ] ,
+        job_modules    =     args["job_modules" ] ,
         job_file       =     "run_chimesmd.cmd")
